Feat: 공유 기능을 홈페이지 부분으로 일부 이동
에디터 페이지에서 공유 문서 불러오는 기능을 홈페이지에서 불러오도록 변경
