+++ client/src/components/editor/toolbar/ToolBarButton.jsx
@@ -6,7 +6,6 @@ import {
   faFileMedical,
   faFileDownload,
   faFileUpload,
-  faFileImport,
   faFileExport,
   faTerminal,
 } from "@fortawesome/free-solid-svg-icons";
@@ -22,7 +21,6 @@ const BUTTON_TYPE = {
   SAVE: faFileDownload,
   LOAD: faFileUpload,
   SHARE: faFileExport,
-  S_LOAD: faFileImport,
   TERMINAL: faTerminal,
 };
 
@@ -68,24 +66,6 @@ const BUTTON_HANDLER = {
       utils.copyText(shareId);
     } else {
       shareDocument();
-    }
-  },
-  S_LOAD: (cellDispatch, cellManager) => {
-    const loadDocument = async (shareId) => {
-      const result = await request.do("SHARE", "GET", null, shareId);
-      const doc = await result.text();
-      if (doc) {
-        cellManager.load(doc);
-      }
-      cellDispatch(cellActionCreator.shareLoadFinish());
-    };
-    const shareId = prompt(
-      "공유 문서의 ID를 입력하세요.",
-      "Input a Document ID"
-    );
-    if (shareId) {
-      cellDispatch(cellActionCreator.shareLoad());
-      loadDocument(shareId);
     }
   },
   TERMINAL: (tmp, temp, terminalDispatch) => {
